# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise AI Business Operations Copilot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  pending: 'pending',
  processing: 'processing',
  analyzed: 'analyzed',
  failed: 'failed',
} as const;

export interface Document {
  id: number;
  name: string;
  fileType: string;
  objectPath: string;
  status: DocumentStatus;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  pageCount?: number | null;
  uploadedAt: string;
  /** @nullable */
  analyzedAt?: string | null;
}

export interface DocumentAnalysis {
  id: number;
  documentId: number;
  /** @nullable */
  executiveSummary?: string | null;
  /**
     * JSON array of KPI objects
     * @nullable
     */
  kpis?: string | null;
  /**
     * JSON array of finding strings
     * @nullable
     */
  keyFindings?: string | null;
  /** @nullable */
  positiveFindings?: string | null;
  /** @nullable */
  negativeFindings?: string | null;
  /** @nullable */
  risks?: string | null;
  /** @nullable */
  opportunities?: string | null;
  /**
     * JSON array of recommendation objects
     * @nullable
     */
  recommendations?: string | null;
  /** @nullable */
  anomalies?: string | null;
  /** @nullable */
  departments?: string | null;
  /** @nullable */
  businessHealthScore?: number | null;
  /** @nullable */
  confidenceScore?: number | null;
  createdAt: string;
}

export type DocumentDetail = Document & ({
  /** @nullable */
  extractedText?: string | null;
  analysis?: DocumentAnalysis;
});

export interface DocumentInput {
  name: string;
  fileType: string;
  objectPath: string;
  /** @nullable */
  fileSize?: number | null;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  generating: 'generating',
  ready: 'ready',
  failed: 'failed',
} as const;

export interface Report {
  id: number;
  title: string;
  status: ReportStatus;
  documentCount?: number;
  /** @nullable */
  confidenceScore?: number | null;
  createdAt: string;
}

export type ReportDetail = Report & ({
  documentIds?: number[];
  /** @nullable */
  executiveSummary?: string | null;
  /** @nullable */
  keyFindings?: string | null;
  /** @nullable */
  kpis?: string | null;
  /** @nullable */
  trends?: string | null;
  /** @nullable */
  positiveFindings?: string | null;
  /** @nullable */
  negativeFindings?: string | null;
  /** @nullable */
  riskAssessment?: string | null;
  /** @nullable */
  opportunities?: string | null;
  /** @nullable */
  strategicRecommendations?: string | null;
  /** @nullable */
  immediateActions?: string | null;
  /** @nullable */
  longTermActions?: string | null;
  /** @nullable */
  financialSummary?: string | null;
  /** @nullable */
  swotAnalysis?: string | null;
  /** @nullable */
  aiInsights?: string | null;
});

export interface ReportInput {
  title: string;
  /** @minItems 1 */
  documentIds: number[];
}

export interface Conversation {
  id: number;
  title: string;
  documentIds?: number[];
  messageCount?: number;
  /** @nullable */
  lastMessageAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export type ConversationDetail = Conversation & {
  messages?: Message[];
};

export interface ConversationInput {
  title: string;
  documentIds?: number[];
}

export interface MessageInput {
  content: string;
}

export interface DashboardSummary {
  totalDocuments: number;
  analyzedDocuments: number;
  pendingDocuments?: number;
  totalReports: number;
  totalConversations: number;
  recentAnalyses: number;
  /** @nullable */
  averageHealthScore?: number | null;
  /** @nullable */
  averageConfidenceScore?: number | null;
  topRisks?: string[];
  topOpportunities?: string[];
  unreadNotifications?: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  document_uploaded: 'document_uploaded',
  document_analyzed: 'document_analyzed',
  report_generated: 'report_generated',
  conversation_started: 'conversation_started',
  notification: 'notification',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  documentId?: number | null;
  /** @nullable */
  reportId?: number | null;
  createdAt: string;
}

export interface KpiDataPoint {
  date: string;
  value: number;
  /** @nullable */
  label?: string | null;
}

export interface KpiTrendSeries {
  name: string;
  data: KpiDataPoint[];
}

export type SearchResultType = typeof SearchResultType[keyof typeof SearchResultType];


export const SearchResultType = {
  document: 'document',
  report: 'report',
  conversation: 'conversation',
  recommendation: 'recommendation',
} as const;

export interface SearchResult {
  type: SearchResultType;
  id: number;
  title: string;
  excerpt: string;
  /** @nullable */
  relevance?: number | null;
  /** @nullable */
  createdAt?: string | null;
}

export interface SearchResults {
  query: string;
  results: SearchResult[];
  total: number;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  risk_detected: 'risk_detected',
  kpi_alert: 'kpi_alert',
  analysis_complete: 'analysis_complete',
  report_ready: 'report_ready',
  anomaly_detected: 'anomaly_detected',
} as const;

export interface Notification {
  id: number;
  type: NotificationType;
  message: string;
  read: boolean;
  /** @nullable */
  documentId?: number | null;
  /** @nullable */
  reportId?: number | null;
  createdAt: string;
}

export type ListDocumentsParams = {
status?: ListDocumentsStatus;
search?: string;
};

export type ListDocumentsStatus = typeof ListDocumentsStatus[keyof typeof ListDocumentsStatus];


export const ListDocumentsStatus = {
  pending: 'pending',
  processing: 'processing',
  analyzed: 'analyzed',
  failed: 'failed',
} as const;

export type GetDashboardActivityParams = {
limit?: number;
};

export type SearchAllParams = {
/**
 * @minLength 1
 */
q: string;
/**
 * Comma-separated entity types to search
 */
types?: string;
};

export type ListNotificationsParams = {
unreadOnly?: boolean;
};

export type MarkAllNotificationsRead200 = {
  updated: number;
};