# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise AI Business Operations Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, processing, analyzed, failed]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Register an uploaded document for analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document with its analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/analyze:
    post:
      operationId: analyzeDocument
      tags: [documents]
      summary: Trigger AI analysis of a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Analysis started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all generated reports
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Generate a new executive report from selected documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "202":
          description: Report generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a full report with all sections
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Document:
      type: object
      required: [id, name, fileType, objectPath, status, uploadedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        fileType:
          type: string
        objectPath:
          type: string
        status:
          type: string
          enum: [pending, processing, analyzed, failed]
        fileSize:
          type: ["integer", "null"]
        pageCount:
          type: ["integer", "null"]
        uploadedAt:
          type: string
          format: date-time
        analyzedAt:
          type: ["string", "null"]
          format: date-time
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/Document"
        - type: object
          properties:
            extractedText:
              type: ["string", "null"]
            analysis:
              $ref: "#/components/schemas/DocumentAnalysis"
    DocumentInput:
      type: object
      required: [name, fileType, objectPath]
      properties:
        name:
          type: string
        fileType:
          type: string
        objectPath:
          type: string
        fileSize:
          type: ["integer", "null"]
    DocumentAnalysis:
      type: object
      required: [id, documentId, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        executiveSummary:
          type: ["string", "null"]
        kpis:
          type: ["string", "null"]
          description: JSON array of KPI objects
        keyFindings:
          type: ["string", "null"]
          description: JSON array of finding strings
        positiveFindings:
          type: ["string", "null"]
        negativeFindings:
          type: ["string", "null"]
        risks:
          type: ["string", "null"]
        opportunities:
          type: ["string", "null"]
        recommendations:
          type: ["string", "null"]
          description: JSON array of recommendation objects
        anomalies:
          type: ["string", "null"]
        departments:
          type: ["string", "null"]
        businessHealthScore:
          type: ["number", "null"]
        confidenceScore:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    Report:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [generating, ready, failed]
        documentCount:
          type: integer
        confidenceScore:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    ReportDetail:
      allOf:
        - $ref: "#/components/schemas/Report"
        - type: object
          properties:
            documentIds:
              type: array
              items:
                type: integer
            executiveSummary:
              type: ["string", "null"]
            keyFindings:
              type: ["string", "null"]
            kpis:
              type: ["string", "null"]
            trends:
              type: ["string", "null"]
            positiveFindings:
              type: ["string", "null"]
            negativeFindings:
              type: ["string", "null"]
            riskAssessment:
              type: ["string", "null"]
            opportunities:
              type: ["string", "null"]
            strategicRecommendations:
              type: ["string", "null"]
            immediateActions:
              type: ["string", "null"]
            longTermActions:
              type: ["string", "null"]
            financialSummary:
              type: ["string", "null"]
            swotAnalysis:
              type: ["string", "null"]
            aiInsights:
              type: ["string", "null"]
    ReportInput:
      type: object
      required: [title, documentIds]
      properties:
        title:
          type: string
        documentIds:
          type: array
          items:
            type: integer
          minItems: 1
...
    Notification:
      type: object
      required: [id, type, message, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [risk_detected, kpi_alert, analysis_complete, report_ready, anomaly_detected]
        message:
          type: string
        read:
          type: boolean
        documentId:
          type: ["integer", "null"]
        reportId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time